Remove wayland socket to prevent game list background flickering
After upgrading from the unofficial flatpak to the official flatpak, the list of games now flickers after loading a savestate in a game (Crash Bandicoot in this case), behind the game itself. I fixed this locally by disabling the wayland socket via flatseal, forcing the emulator to run in Xwayland. This is on KDE wayland session.